#255: Began refactoring out of carousel
