+++ src/components/Carousel/utils.ts
@@ -1,0 +1,6 @@
+export const noop = () => {};
+
+export const defaultStatusFormatter = (current: number, total: number) => `${current} of ${total}`;
+
+export const isKeyboardEvent = (e?: React.MouseEvent | React.KeyboardEvent): e is React.KeyboardEvent =>
+    e ? e.hasOwnProperty('key') : false;
